requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pillow>=10.0.0
opencv-python>=4.8.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
lxml>=4.9.0
html5lib>=1.1
google-genai>=0.1.0 